Initial LookML conversion tool (#16)
This PR is a first pass at implementing a tool to convert LookML to dbt SL definitions. An example usage of this tool can be found here (internal to dbt Labs).

Since this is my first meaningful PR in this repo, I felt that doing some refactoring would help me understand it better. In hindsight, I should have done that as a separate PR, so I'm sorry about that.

I will be sure to squash these commits before I merge.